% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corHeatmap.R
\name{corHeatmap}
\alias{corHeatmap}
\alias{corHeatmapOutput}
\alias{corHeatmapUI}
\title{Shiny correlation-based heatmap}
\usage{
corHeatmap(input, output, session)

corHeatmapUI(id)

corHeatmapOutput(id)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{id}{session identifier}
}
\description{
Shiny correlation-based heatmap
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

