% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_cor.R
\name{dist_cor}
\alias{dist_cor}
\title{Correlation-based distance}
\usage{
dist_cor(x, beta = 1)
}
\arguments{
\item{x}{matrix to compute correlation}

\item{beta}{exponent to attenuate correlation}
}
\description{
Correlation-based distance
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

