% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_reorder.R
\name{heatmap_reorder}
\alias{heatmap_reorder}
\title{Reorder heatmap}
\usage{
heatmap_reorder(x, distfun)
}
\arguments{
\item{x}{matrix of values}

\item{distfun}{distance function for \code{x}}
}
\description{
Reorder heatmap
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

