% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapApp.R
\name{heatmapApp}
\alias{heatmapApp}
\title{Correlation-based heatmap app}
\usage{
heatmapApp()
}
\description{
Runs Shiny app for correlation based heatmap, built on d3heatmap.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

