% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_pval.R
\name{lm_pval}
\alias{lm_pval}
\title{anova p-values grouped by column}
\usage{
lm_pval(object, form, group = NULL, type = c("drop1", "anova"))
}
\arguments{
\item{object}{data.frame or tbl_df}

\item{form}{formula}

\item{group}{column to group by (or not if \code{NULL})}

\item{type}{type of p-values (one of c("drop1","anova"))}
}
\description{
anova p-values grouped by column
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

