% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_resid.R
\name{lm_resid}
\alias{lm_resid}
\title{adjust values based on factors}
\usage{
lm_resid(object, form, center = TRUE, group = NULL)
}
\arguments{
\item{object}{data.frame or tbl_df}

\item{form}{formula}

\item{center}{center on mean if \code{TRUE}}

\item{group}{column to group by (or not if \code{NULL})}
}
\description{
adjust values based on factors
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

