% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormap.R
\name{corheatmap}
\alias{corheatmap}
\alias{plot.corheatmap}
\title{Correlation-based heatmap}
\usage{
corheatmap(df, row_names = is.null(rownames(df)), cluster = TRUE,
  beta = 1, num_rows = 500)

\method{plot}{corheatmap}(x, rowname = TRUE, colname = TRUE,
  cluster = attr(x, "cluster"), beta = attr(x, "beta"),
  category = c("div", "seq", "qual"), d3map = interactive(),
  palette = palettes, ...)
}
\arguments{
\item{df}{data frame with matrix to be plotted}

\item{row_names}{column number or name with row names, or vector of row names (default column 1 if no rownames on \code{df})}

\item{cluster}{cluster rows and columns if \code{TRUE} (default)}

\item{beta}{attenuation parameter for clustering (default \code{1})}

\item{num_rows}{number of rows to plot}

\item{rowname}{include row names in plot if \code{TRUE}}

\item{colname}{include column names in plot if \code{TRUE}}

\item{category}{category of \code{RColorBrewer} palettes (see \code{\link[RColorBrewer]{brewer.pal.info}})}

\item{d3map}{use d3heatmap interactive plot if \code{TRUE}}

\item{palette}{colorblind friendly palette to plot (see \code{\link[RColorBrewer]{brewer.pal.info}})}

\item{...}{other plot parameters (ignored)}

\item{cluster}{cluster rows and columns if \code{TRUE} (default)}

\item{beta}{attenuation parameter for clustering (default \code{1})}
}
\description{
Correlation-based heatmap
}
\examples{
corheatmap(mtcars)

plot(corheatmap(mtcars), d3map=FALSE)

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

